RF: move simple deduplication to its own app.
