+++ disclosure/settings.py
@@ -43,6 +43,7 @@ INSTALLED_APPS = (
     'calaccess_raw',
     'netfile_raw',
     'zipcode_metro_raw',
+    'text_dedupe',
     'locality',  # dep: none
     'ballot',  # dep: locality
     'finance',  # dep: locality, ballot
