+++ locality/models.py
@@ -1,11 +1,8 @@
-import numpy as np
-
-from django.core.exceptions import MultipleObjectsReturned, ObjectDoesNotExist
-from django.db import models, transaction
+from django.db import models
 from django.db.models.fields.related import OneToOneRel
-from django.db.models.signals import post_save
-from django.dispatch import receiver
 from django.utils.encoding import python_2_unicode_compatible
+
+from text_dedupe import dedupe, DedupeMixin
 
 
 class ReverseLookupStringMixin(object):
@@ -25,72 +22,15 @@ class ReverseLookupStringMixin(object):
         return unicode(obj) if obj else ''
 
 
-class LocalityQuerySet(models.QuerySet):
-    def get(self, *args, **kwargs):
-        if np.any([key.startswith('name__') for key in kwargs]):
-            raise NotImplementedError("name__ attributes.")
-
-        try:
-            return super(LocalityQuerySet, self).get(*args, **kwargs)
-        except ObjectDoesNotExist as odne:
-            # Already did our best; stick with the exception.
-            if 'name' not in kwargs or np.any([key.startswith('aliases') in kwargs]):
-                raise
-
-        # Try augmented search
-        name = kwargs.pop('name')
-        query = models.Q(aliases__contains=',%s,' % name)
-        query |= models.Q(aliases__startswith='%s,' % name)
-        query |= models.Q(aliases__endswith=',%s' % name)
-        args += (query,)
-        try:
-            return super(LocalityQuerySet, self).get(*args, **kwargs)
-        except MultipleObjectsReturned:
-            raise odne  # go with original error; clearer to users.
-
-
-class LocalityAliasManager(models.Manager):
-    def get_queryset(self):
-        return LocalityQuerySet(self.model, using=self._db)
-
-
 @python_2_unicode_compatible
-class Locality(models.Model, ReverseLookupStringMixin):
+@dedupe('name')
+class Locality(DedupeMixin, ReverseLookupStringMixin):
     """
     A base table that gives a globally unique ID to any
     location (city, state, etc)
     """
     name = models.CharField(max_length=128, blank=True, null=True, default=None)
     short_name = models.CharField(max_length=32, blank=True, null=True, default=None)
-    aliases = models.CharField(max_length=4096, blank=True, null=True, default=None,
-                               help_text="Comma-separated list of alternate spellings for `name`")
-
-    objects = models.Manager()
-    aliased_objects = LocalityAliasManager()
-
-    @classmethod
-    def name_is_aliased(cls, name):
-        if name is None:
-            return False
-        for model in cls.objects.filter(aliases__contains=name):
-            if name in model.aliases.split(','):
-                return True
-        return False
-
-    def clean(self):
-        """Make aliases unique & clean."""
-        if self.aliases is not None:
-            aliases_set = set([a.strip()
-                               for a in self.aliases.strip().split(',')])
-            aliases_cleaned = ','.join(aliases_set)
-            self.aliases = aliases_cleaned or None  # convert blank to None
-        return super(Locality, self).clean()
-
-    def save(self, *args, **kwargs):
-        self.clean()
-        if self.name_is_aliased(self.name):
-            raise Exception('No fair saving an aliased name!')
-        super(Locality, self).save(*args, **kwargs)
 
     def __str__(self):
         return (ReverseLookupStringMixin.__str__(self) or
@@ -112,15 +52,13 @@ class FipsMixin(Locality):
 
 
 @python_2_unicode_compatible
+@dedupe('name')
 class City(FipsMixin):
     """
     City
     """
     county = models.ForeignKey('County', blank=True, null=True, default=None)
     state = models.ForeignKey('State')
-
-    objects = models.Manager()
-    aliased_objects = LocalityAliasManager()
 
     def __str__(self):
         return '%s, %s' % (self.name or self.short_name, self.state)
@@ -137,9 +75,6 @@ class County(FipsMixin):
     """
     state = models.ForeignKey('State')
 
-    objects = models.Manager()
-    aliased_objects = LocalityAliasManager()
-
     def __str__(self):
         # See https://code.djangoproject.com/ticket/25218 on why __unicode__
         return '%s, %s' % (Locality.__unicode__(self), self.state)
@@ -154,8 +89,6 @@ class State(FipsMixin):
     """
     State
     """
-    # Uses short_name, so don't use the alias manager.
-
     def __str__(self):
         return self.short_name or self.name
 
@@ -171,52 +104,9 @@ class ZipCode(Locality):
     city = models.ForeignKey('City', blank=True, null=True, default=None)
     county = models.ForeignKey('County', blank=True, null=True, default=None)
     state = models.ForeignKey('State', blank=True, null=True, default=None)
-    # Uses short_name, so don't use the alias manager.
 
     def __str__(self):
         return self.short_name or self.name
-
-
-@receiver(post_save, sender=Locality)
-@receiver(post_save, sender=City)
-@receiver(post_save, sender=County)
-@transaction.atomic
-def update_city_aliases(sender, instance, **kwargs):
-    """Merge Locality models on obj1.name in obj2.aliases"""
-    if instance.aliases:
-        current_aliases = []
-        new_aliases = instance.aliases.split(',')
-
-        # Resolve which models have a name that is
-        # now an alias.
-        remote_models = [obj
-                         for a in new_aliases
-                         for obj in sender.objects.filter(name__iexact=a)]
-        if len(remote_models) == 1 and remote_models[0].aliases:
-            # Remote model already has aliases, so flip things:
-            # Call that the instance, and make the instance
-            #   one of the remote models
-            new_aliases = [instance.name]
-            instance, remote_models = remote_models[0], [instance]
-            current_aliases = instance.aliases.split(',')
-
-        # Migrate any foreign keys off of aliases, then delete.
-        for model in remote_models:
-            for relationship in model._meta.related_objects:
-                attr = relationship.name
-                if hasattr(model, attr):
-                    for obj in getattr(model, attr).get_queryset():
-                        setattr(obj, relationship.field.name, instance)
-            model.delete()  # Remove the (now obsolete) aliased model.
-
-        # Normalize aliases
-        all_aliases = current_aliases + new_aliases
-        all_aliases_text = ','.join(all_aliases)
-        if instance.aliases != all_aliases_text:
-            instance.aliases = all_aliases_text
-            instance.save()
-
-    return instance
 
 
 class AddressMixin(models.Model):
