+++ text_dedupe/__init__.py
@@ -1,0 +1,2 @@
+from .models import dedupe, DedupeMixin
+__all__ = ['dedupe', 'DedupeMixin']
